GENIE3/GENIE/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
decorator==5.1.1
dj-database-url==1.0.0
Django==4.0.5
django-cors-headers==3.13.0
django-database-url==1.0.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-writable-nested==0.7.0
dynaconf==3.1.9
et-xmlfile==1.1.0
gunicorn==20.1.0
hvac==0.11.2
idna==3.3
knox==0.1.14
loguru==0.6.0
numpy==1.22.4
openai==0.20.0
openpyxl==3.0.10
pandas==1.4.2
pandas-stubs==1.2.0.61
permission==0.4.1
psycopg2==2.9.3
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
shortuuid==1.0.9
six==1.16.0
sqlparse==0.4.2
tqdm==4.64.0
tzdata==2022.1
urllib3==1.26.9
validators==0.20.0
whitenoise==6.2.0
win32-setctime==1.1.0
